requirements.txt
pandas
numpy
python-constraint
together